requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
aiosqlite==0.20.0
aioyoomoney-api==0.0.9
attrs==23.2.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
disnake==2.9.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyNaCl==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
